Rename 'gcloud/_apitools' -> 'gcloud/streaming'.
Works around a bug in 'nose', which forcibly ignores subpackages with
leading '_'.

See:
http://stackoverflow.com/questions/23615924
https://github.com/nose-devs/nose/blob/master/nose/loader.py#L169.
