+++ gcloud/storage/test_blob.py
@@ -409,8 +409,8 @@ class Test_Blob(unittest2.TestCase):
         from six.moves.urllib.parse import urlsplit
         from tempfile import NamedTemporaryFile
         from gcloud._testing import _Monkey
-        from gcloud._apitools import http_wrapper
-        from gcloud._apitools import transfer
+        from gcloud.streaming import http_wrapper
+        from gcloud.streaming import transfer
         BLOB_NAME = 'blob-name'
         UPLOAD_URL = 'http://example.com/upload/name/key'
         DATA = b'ABCDEF'
@@ -470,7 +470,7 @@ class Test_Blob(unittest2.TestCase):
         from six.moves.urllib.parse import parse_qsl
         from six.moves.urllib.parse import urlsplit
         from tempfile import NamedTemporaryFile
-        from gcloud._apitools import http_wrapper
+        from gcloud.streaming import http_wrapper
         BLOB_NAME = 'parent/child'
         UPLOAD_URL = 'http://example.com/upload/name/parent%2Fchild'
         DATA = b'ABCDEF'
@@ -518,7 +518,7 @@ class Test_Blob(unittest2.TestCase):
         from six.moves.urllib.parse import parse_qsl
         from six.moves.urllib.parse import urlsplit
         from tempfile import NamedTemporaryFile
-        from gcloud._apitools import http_wrapper
+        from gcloud.streaming import http_wrapper
         BLOB_NAME = 'blob-name'
         UPLOAD_URL = 'http://example.com/upload/name/key'
         DATA = b'ABCDEF'
@@ -584,7 +584,7 @@ class Test_Blob(unittest2.TestCase):
         from six.moves.http_client import OK
         from six.moves.urllib.parse import parse_qsl
         from six.moves.urllib.parse import urlsplit
-        from gcloud._apitools import http_wrapper
+        from gcloud.streaming import http_wrapper
         BLOB_NAME = 'blob-name'
         UPLOAD_URL = 'http://example.com/upload/name/key'
         DATA = b'ABCDEF'
@@ -623,7 +623,7 @@ class Test_Blob(unittest2.TestCase):
         from six.moves.http_client import OK
         from six.moves.urllib.parse import parse_qsl
         from six.moves.urllib.parse import urlsplit
-        from gcloud._apitools import http_wrapper
+        from gcloud.streaming import http_wrapper
         BLOB_NAME = 'blob-name'
         UPLOAD_URL = 'http://example.com/upload/name/key'
         DATA = u'ABCDEF\u1234'
