+++ gcloud/streaming/exceptions.py
