+++ tox.ini
@@ -55,7 +55,7 @@ deps = {[testenv:docs]deps}
 passenv = {[testenv:docs]passenv}
 
 [pep8]
-exclude = gcloud/datastore/_datastore_v1_pb2.py,gcloud/bigtable/_generated/*,docs/conf.py,gcloud/_apitools/*
+exclude = gcloud/datastore/_datastore_v1_pb2.py,gcloud/bigtable/_generated/*,docs/conf.py,gcloud/streaming/*
 verbose = 1
 
 [testenv:lint]
